perf(cli): keep the env-graph barrel out of the CLI entry point
Splitting the commands off left a 559KB shared chunk still loaded on every
invocation. That chunk was env-graph + the @env-spec parser + clack prompts
+ ansis, merged together because the entry reached the env-graph barrel
through three edges:

- helpers/error-checks.ts imports { EnvGraph, FileBasedDataSource } from the
  barrel, but the entry only ever wanted InvalidEnvError from that module.
  Moved that class to its own leaf module and pointed the entry at it.
- lib/formatting.ts imported ConfigItem and VarlockError as values, though
  both are only used as parameter annotations. Made it an `import type`, so
  the edge is erased at compile time.
- helpers/telemetry-usage-context.ts already deep-imports env-graph/lib/errors,
  which is a leaf (its only import is my-dash), so it costs nothing.

error-checks.ts re-exports InvalidEnvError, so no other caller changes. The
class is still defined exactly once in the bundle, which is what keeps the
entry's `instanceof InvalidEnvError` catch working.

Entry static closure: 1029KB -> 320KB.
