+++ packages/varlock/src/lib/formatting.ts
@@ -1,7 +1,7 @@
 import ansis, { type AnsiColors, type AnsiStyles } from 'ansis';
 import _ from '@env-spec/utils/my-dash';
 
-import { ConfigItem, VarlockError } from '../env-graph';
+import type { ConfigItem, VarlockError } from '../env-graph';
 import { redactString } from '../runtime/lib/redaction';
 
 type ColorMod = AnsiStyles | AnsiColors;
